courses/data-engineering/kubeflow-examples/pipelines/azurepipeline/code/register/requirements.txt
pathlib2
requests
azureml-sdk
